running: Remove full-upgrade advice
While what it said is technically true, mentioning it will cause people
to (regularly) use it and thereby usually *break* their system.

Plebian ships with Bookworm which will become Stable in 1.5 months or so
and on a Debian Stable system, `apt upgrade` is all you need.

There is 1 exception to that and that is when you upgrade from one
Stable release to the next and in the case of Plebian that would be
upgrading from Bookworm to Trixie.
And then the advice becomes: read and follow the (Trixie) release notes
which will describe how to upgrade your system and that will (highly
likely) include a `full-upgrade` command.

The `full-upgrade` command is actually forcing things and therefor more
often breaks people's system then fixing an issue. Consequently I'm on
some kind of 'mission' to stamp out the use and recommendation of the
`full-upgrade` command and it looks like I'm no longer the only one :-)

The `apt` program has already adapted several things which have been
present in `aptitude` for ages and I hope it'll also adopt the commands
and its *workings* of `safe-upgrade` and `full-upgrade`.